[GPU-Plugin] Integration of a faster version of grow_gpu plugin into mainstream (#2360)
* Integrating a faster version of grow_gpu plugin
1. Removed the older files to reduce duplication
2. Moved all of the grow_gpu files under 'exact' folder
3. All of them are inside 'exact' namespace to avoid any conflicts
4. Fixed a bug in benchmark.py while running only 'grow_gpu' plugin
5. Added cub and googletest submodules to ease integration and unit-testing
6. Updates to CMakeLists.txt to directly build cuda objects into libxgboost

* Added support for building gpu plugins through make flow
1. updated makefile and config.mk to add right targets
2. added unit-tests for gpu exact plugin code

* 1. Added support for building gpu plugin using 'make' flow as well
2. Updated instructions for building and testing gpu plugin

* Fix travis-ci errors for PR#2360
1. lint errors on unit-tests
2. removed googletest, instead depended upon dmlc-core provide gtest cache

* Some more fixes to travis-ci lint failures PR#2360

* Added Rory's copyrights to the files containing code from both.

* updated copyright statement as per Rory's request

* moved the static datasets into a script to generate them at runtime

* 1. memory usage print when silent=0
2. tests/ and test/ folder organization
3. removal of the dependency of googletest for just building xgboost
4. coding style updates for .cuh as well

* Fixes for compilation warnings

* add cuda object files as well when JVM_BINDINGS=ON

+++ tests/cpp/objective/test_regression_obj.cc
@@ -49,7 +49,7 @@ TEST(Objective, LogisticRegressionBasic) {
   std::vector<xgboost::bst_float> preds = {0, 0.1, 0.5, 0.9, 1};
   std::vector<xgboost::bst_float> out_preds = {0.5, 0.524, 0.622, 0.710, 0.731};
   obj->PredTransform(&preds);
-  for (int i = 0; i < preds.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(preds.size()); ++i) {
     EXPECT_NEAR(preds[i], out_preds[i], 0.01);
   }
 }
@@ -97,7 +97,7 @@ TEST(Objective, PoissonRegressionBasic) {
   std::vector<xgboost::bst_float> preds = {0, 0.1, 0.5, 0.9, 1};
   std::vector<xgboost::bst_float> out_preds = {1, 1.10, 1.64, 2.45, 2.71};
   obj->PredTransform(&preds);
-  for (int i = 0; i < preds.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(preds.size()); ++i) {
     EXPECT_NEAR(preds[i], out_preds[i], 0.01);
   }
 }
@@ -132,7 +132,7 @@ TEST(Objective, GammaRegressionBasic) {
   std::vector<xgboost::bst_float> preds = {0, 0.1, 0.5, 0.9, 1};
   std::vector<xgboost::bst_float> out_preds = {1, 1.10, 1.64, 2.45, 2.71};
   obj->PredTransform(&preds);
-  for (int i = 0; i < preds.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(preds.size()); ++i) {
     EXPECT_NEAR(preds[i], out_preds[i], 0.01);
   }
 }
@@ -168,7 +168,7 @@ TEST(Objective, TweedieRegressionBasic) {
   std::vector<xgboost::bst_float> preds = {0, 0.1, 0.5, 0.9, 1};
   std::vector<xgboost::bst_float> out_preds = {1, 1.10, 1.64, 2.45, 2.71};
   obj->PredTransform(&preds);
-  for (int i = 0; i < preds.size(); ++i) {
+  for (int i = 0; i < static_cast<int>(preds.size()); ++i) {
     EXPECT_NEAR(preds[i], out_preds[i], 0.01);
   }
 }
